One more (int) cast.
git-svn-id: svn://vcs.exim.org/pcre/code/trunk@533 2f5784b3-3f2a-0410-8824-cb99058d5e15

+++ pcre_compile.c
@@ -4757,7 +4757,7 @@ for (;; ptr++)
         arg = ++ptr;
         while ((cd->ctypes[*ptr] & (ctype_letter|ctype_digit)) != 0
           || *ptr == '_') ptr++;
-        arglen = ptr - arg;
+        arglen = (int)(ptr - arg);
         }
 
       if (*ptr != CHAR_RIGHT_PARENTHESIS)
